services/api/requirements.txt
flask==3.0.3
redis==5.0.8
prometheus-client==0.20.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0